# sites/unep-dashboard/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UNEP PUBLIC API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/public/questions": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Question list.",
                "description": "Show all questions.<br/>\n                <ul>\n                    <li>\n                        <i>parent_id</i>, define a relation of parent and children between question.\n                    </li>\n                    <li>\n                        <i>value_id</i>, define a relation between question and value/option.\n                    </li>\n                    <li>\n                        <i>type</i>, define question type.\n                    </li>\n                </ul>\n            ",
                "operationId": "getPublicQuestions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "parent_id": null,
                                        "value_id": null,
                                        "code": "43374797",
                                        "name": "Your name",
                                        "type": "FreeText"
                                    },
                                    {
                                        "id": 2,
                                        "parent_id": 1,
                                        "value_id": null,
                                        "code": "43374799",
                                        "name": "Your role",
                                        "type": "FreeText"
                                    }
                                ]
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "content": {
                            "text/plain": {
                                "example": "No Content"
                            }
                        }
                    }
                }
            }
        },
        "/api/public/groups": {
            "get": {
                "tags": [
                    "Country Group"
                ],
                "summary": "Country group list.",
                "description": "Show all country group with countries.<br/>\n                <ul>\n                    <li>\n                        <i>parent_id</i>, define a relation of parent and children between country group.\n                    </li>\n                </ul>\n            ",
                "operationId": "getPublicGroups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "parent_id": null,
                                        "name": "COBSEA",
                                        "code": "COBSEA",
                                        "countries": [
                                            {
                                                "id": 1,
                                                "group_id": 1,
                                                "country": {
                                                    "id": 77,
                                                    "name": "Indonesia",
                                                    "code": "IDN"
                                                }
                                            },
                                            {
                                                "id": 2,
                                                "group_id": 1,
                                                "country": {
                                                    "id": 124,
                                                    "name": "Malaysia",
                                                    "code": "MYS"
                                                }
                                            }
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "content": {
                            "text/plain": {
                                "example": "No Content"
                            }
                        }
                    }
                }
            }
        },
        "/api/public/countries": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Country list.",
                "description": "Show all country with group.",
                "operationId": "getPublicCountries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Afghanistan",
                                        "code": "AFG",
                                        "groups": [
                                            {
                                                "id": 12,
                                                "parent_id": null,
                                                "name": "South Asia Seas",
                                                "code": "SAS",
                                                "pivot": {
                                                    "country_id": 1,
                                                    "group_id": 12
                                                }
                                            },
                                            {
                                                "id": 20,
                                                "parent_id": null,
                                                "name": "Least Developed Countries",
                                                "code": "LDC",
                                                "pivot": {
                                                    "country_id": 1,
                                                    "group_id": 20
                                                }
                                            }
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "content": {
                            "text/plain": {
                                "example": "No Content"
                            }
                        }
                    }
                }
            }
        },
        "/api/public/filters": {
            "get": {
                "tags": [
                    "Filter"
                ],
                "summary": "Filter list.",
                "description": "Show all filters.<br/>\n                <ul>\n                    <li>\n                        <i>parent_id</i>, define a relation of parent and children between filter/indicator.\n                    </li>\n                </ul>\n            ",
                "operationId": "getPublicFilters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "parent_id": null,
                                        "code": "43375025",
                                        "name": "Reporting",
                                        "childrens": [
                                            {
                                                "id": 2,
                                                "parent_id": 1,
                                                "code": "105885247",
                                                "name": "As an individual",
                                                "childrens": [],
                                                "locale": []
                                            },
                                            {
                                                "id": 3,
                                                "parent_id": 1,
                                                "code": "105885313",
                                                "name": "On behalf of an organisation",
                                                "childrens": [],
                                                "locale": []
                                            },
                                            {
                                                "id": 4,
                                                "parent_id": 1,
                                                "code": "106054616",
                                                "name": "Other",
                                                "childrens": [],
                                                "locale": []
                                            }
                                        ],
                                        "locale": []
                                    }
                                ]
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "content": {
                            "text/plain": {
                                "example": "No Content"
                            }
                        }
                    }
                }
            }
        },
...
    }
}